Add some spelling correction
I normally don't think machine-based spelling detection is worth it, because for
me, it almost always flags a million false positives and very rarely catch any
real typos.

...until I got this ThinkPad. For some reason I do a million transposition typos
with the keyboard on this thing. So while I get used to the right amount of
actuation force, I might actually benefit from a machine comparing my words to a
dictionary.