try removing csharp-mode configuration entirely
csharp-mode is included in modern versions of Emacs, but also these settings
were – I think – mainly used to work around bugs in old versions of csharp-mode
as well as to comply with style rules at my previous employer.

I'll try removing this block entirely and then if I feel the need to I can
reintroduce parts of it as I need to.